+++ Sources/KeyboardKit/Settings/KeyboardSettingsLink.swift
@@ -8,28 +8,24 @@
 
 import SwiftUI
 
-/**
- This link can be used to navigate a user to System Settings.
-
- The link's behavior is a bit inconsistent. It should always
- link to an app's custom settings in System Settings, but it
- can sometimes just link to the System Settings root instead.
- The reason for this behavior is unknown.
- 
- To improve the behavior of this link, add an empty Settings
- Bundle to your application bundle. This makes your app more
- likely to open the correct screen.
- */
+/// This link can be used to navigate to System Settings.
+///
+/// The link's behavior can be a bit inconsistent. It should
+/// always link to an app's settings in System Settings, but
+/// can sometimes just link to the root. The reason for this
+/// is currently unknown.
+///
+/// To improve the behavior of the link, simply add an empty
+/// Settings Bundle to your app. This will make the app much
+/// more likely to open the correct screen.
 public struct KeyboardSettingsLink<Content: View>: View {
 
-    /**
-     Create a keyboard settings link with a custom label.
-
-     - Parameters:
-       - url: The url to navigate to, by default `.keyboardSettings`.
-       - addNavigationArrow: Whether or not to add a trailing navigation disclosure arrow, by default `false`.
-       - label: The link label.
-     */
+    /// Create a settings link with a custom label.
+    ///
+    /// - Parameters:
+    ///   - url: The url to navigate to, by default ``Foundation/URL/keyboardSettings``.
+    ///   - addNavigationArrow: Whether to add a trailing disclosure arrow, by default `false`.
+    ///   - label: A label view builder.
     public init(
         url: URL? = .keyboardSettings,
         addNavigationArrow: Bool = false,
@@ -55,14 +51,12 @@ public struct KeyboardSettingsLink<Content: View>: View {
 
 public extension KeyboardSettingsLink where Content == Label<Text, Image> {
 
-    /**
-     Create a keyboard settings link with a title and icon.
-
-     - Parameters:
-       - title: The button text, by default `System Settings`.
-       - icon: The button icon, by default `.keyboardSettings`.
-       - url: The url to navigate to, by default `.keyboardSettings`.
-     */
+    /// Create a settings link with a title and icon.
+    ///
+    /// - Parameters:
+    ///   - title: The button text, by default `System Settings`.
+    ///   - icon: The button icon, by default ``SwiftUI/Image/keyboardSettings``.
+    ///   - url: The url to navigate to, by default ``Foundation/URL/keyboardSettings``.
     init(
         title: String = "System Settings",
         icon: Image = .keyboardSettings,