+++ Sources/KeyboardKit/Settings/Url+KeyboardSettings.swift
@@ -14,23 +14,21 @@ import UIKit
 
 public extension URL {
     
-    /**
-     Use this URL to deep link to your app's custom settings
-     in System Settings.
-     
-     The URL behavior is inconsistent. It should always link
-     to an app's custom settings in System Settings, but can
-     sometimes just link to the System Settings root instead.
-     The reason for this behavior is unknown.
-     
-     To improve the behavior of this link, just add an empty
-     Settings Bundle to your application bundle. This causes
-     your app to more consistently open the correct screen.
-     
-     Maybe `appSettings` would be a better name for this URL,
-     but this name was selected to reduce the risk of naming
-     conflicts with other URL extensions.
-     */
+    /// This URL will deep link to the app's custom settings
+    /// in System Settings.
+    ///
+    /// The URL's behavior is inconsistent. It should always
+    /// link to the app's custom settings in System Settings,
+    /// but can sometimes just link to the root instead. The
+    /// reason for this behavior is currently unknown.
+    ///
+    /// To improve this link behavior, add an empty Settings
+    /// Bundle to your app. This will causes the app to more
+    /// consistently open the correct System Settings screen.
+    ///
+    /// Maybe `appSettings` would be a better name, but this
+    /// name was used to reduce the risk of naming conflicts
+    /// with other URL extensions.
     static var keyboardSettings: URL? {
         #if os(iOS) || os(tvOS)
         URL(string: UIApplication.openSettingsURLString)