+++ Sources/KeyboardKit/State/KeyboardStateLabel.swift
@@ -8,27 +8,38 @@
 
 import SwiftUI
 
-/**
- This label can be used to show different views depending on
- a provided state.
-
- While you can use this label with any boolean, the intended
- use is to show if the keyboard extension is enabled, if the
- app has full access etc.
- */
+/// This label can be used to show different views depending
+/// on a certain provided state.
+///
+/// While you can use this label with any bool, the intended
+/// use is to show if a keyboard extension is enabled, if is
+/// has full access etc.
 public struct KeyboardStateLabel<EnabledIcon: View, DisabledIcon: View>: View {
-
-    /**
-     Create an enabled label.
-
-     - Parameters:
-       - isEnabled: Whether or not the state is enabled.
-       - enabledText: The text to show when the state is enabled.
-       - enabledIcon: The icon to show when the state is enabled, by default `checkmark`.
-       - disabledText: The text to show when the state is disabled.
-       - disabledIcon: The icon to show when the state is disabled, by default `exclamationmark.triangle`.
-       - style: The style to use, by default ``KeyboardStateLabelStyle/standard``.
-     */
+    
+    /// Create an enabled label.
+    ///
+    /// - Parameters:
+    ///   - isEnabled: Whether or not the state is enabled.
+    ///   - enabledText: The text to show when the state is enabled.
+    ///   - enabledIcon: The icon to show when the state is enabled, by default `checkmark`.
+    ///   - disabledText: The text to show when the state is disabled.
+    ///   - disabledIcon: The icon to show when the state is disabled, by default `exclamationmark.triangle`.
+    public init(
+        isEnabled: Bool,
+        enabledIcon: EnabledIcon,
+        enabledText: String,
+        disabledIcon: DisabledIcon,
+        disabledText: String
+    ) {
+        self.isEnabled = isEnabled
+        self.enabledText = enabledText
+        self.enabledIcon = enabledIcon
+        self.disabledText = disabledText
+        self.disabledIcon = disabledIcon
+        self.initStyle = nil
+    }
+    
+    @available(*, deprecated, message: "Style this view with .keyboardStateLabelStyle instead.")
     public init(
         isEnabled: Bool,
         enabledIcon: EnabledIcon,
@@ -42,7 +53,7 @@ public struct KeyboardStateLabel<EnabledIcon: View, DisabledIcon: View>: View {
         self.enabledIcon = enabledIcon
         self.disabledText = disabledText
         self.disabledIcon = disabledIcon
-        self.style = style
+        self.initStyle = style
     }
 
     private let isEnabled: Bool
@@ -50,7 +61,9 @@ public struct KeyboardStateLabel<EnabledIcon: View, DisabledIcon: View>: View {
     private let enabledIcon: EnabledIcon
     private let disabledText: String
     private let disabledIcon: DisabledIcon
-    private let style: KeyboardStateLabelStyle
+    
+    @Environment(\.keyboardStateLabelStyle)
+    private var envStyle
 
     public var body: some View {
         Label {
@@ -59,21 +72,40 @@ public struct KeyboardStateLabel<EnabledIcon: View, DisabledIcon: View>: View {
             icon.foregroundColor(iconColor).font(iconFont)
         }
     }
+    
+    // MARK: - Deprecated
+    
+    private typealias Style = KeyboardStateLabelStyle
+    private let initStyle: Style?
+    private var style: Style { initStyle ?? envStyle }
 }
 
 public extension KeyboardStateLabel where EnabledIcon == Image, DisabledIcon == Image {
-
-    /**
-     Create an enabled label.
-
-     - Parameters:
-       - isEnabled: Whether or not the state is enabled.
-       - enabledText: The text to show when the state is enabled.
-       - enabledIcon: The icon to show when the state is enabled, by default `checkmark`.
-       - disabledText: The text to show when the state is disabled.
-       - disabledIcon: The icon to show when the state is disabled, by default `exclamationmark.triangle`.
-       - style: The style to use, by default ``KeyboardStateLabelStyle/standard``.
-     */
+    
+    /// Create an enabled label.
+    ///
+    /// - Parameters:
+    ///   - isEnabled: Whether or not the state is enabled.
+    ///   - enabledText: The text to show when the state is enabled.
+    ///   - enabledIcon: The icon to show when the state is enabled, by default `checkmark`.
+    ///   - disabledText: The text to show when the state is disabled.
+    ///   - disabledIcon: The icon to show when the state is disabled, by default `exclamationmark.triangle`.
+    init(
+        isEnabled: Bool,
+        enabledIcon: Image = .init(systemName: "checkmark"),
+        enabledText: String,
+        disabledIcon: Image = .init(systemName: "exclamationmark.triangle"),
+        disabledText: String
+    ) {
+        self.isEnabled = isEnabled
+        self.enabledText = enabledText
+        self.enabledIcon = enabledIcon
+        self.disabledText = disabledText
+        self.disabledIcon = disabledIcon
+        self.initStyle = nil
+    }
+    
+    @available(*, deprecated, message: "Style this view with .keyboardStateLabelStyle instead.")
     init(
         isEnabled: Bool,
         enabledIcon: Image = .init(systemName: "checkmark"),
@@ -87,7 +119,7 @@ public extension KeyboardStateLabel where EnabledIcon == Image, DisabledIcon == 
         self.enabledIcon = enabledIcon
         self.disabledText = disabledText
         self.disabledIcon = disabledIcon
-        self.style = style
+        self.initStyle = style
     }
 }
 
@@ -127,86 +159,28 @@ private extension KeyboardStateLabel {
     }
 }
 
-/**
- This style can be used with ``KeyboardStateLabel`` views.
- */
-public struct KeyboardStateLabelStyle: Equatable {
+#Preview {
 
-    /**
-     This style can be used with a ``KeyboardStateLabel``.
-     */
-    public struct ComponentStyle: Equatable {
-
-        public init(color: Color?, font: Font?) {
-            self.color = color
-            self.font = font
+    List {
+        Section {
+            ForEach([true, false], id: \.self) {
+                KeyboardStateLabel(
+                    isEnabled: $0,
+                    enabledText: "Enabled",
+                    disabledText: "Disabled"
+                )
+            }
         }
-
-        public var color: Color?
-        public var font: Font?
-    }
-
-    /**
-     Create an enabled label style.
-
-     - Parameters:
-       - enabledIcon: The icon style to use when the state is enabled.
-       - enabledText: The text style to use when the state is enabled.
-       - disabledIcon: The icon style to use when the state is disabled.
-       - disabledText: The text style to use when the state is disabled.
-     */
-    public init(
-        enabledIcon: ComponentStyle = .init(color: .green, font: nil),
-        enabledText: ComponentStyle = .init(color: nil, font: nil),
-        disabledIcon: ComponentStyle = .init(color: .orange, font: nil),
-        disabledText: ComponentStyle = .init(color: nil, font: nil)
-    ) {
-        self.enabledIcon = enabledIcon
-        self.enabledText = enabledText
-        self.disabledIcon = disabledIcon
-        self.disabledText = disabledText
-    }
-
-    /// The icon style to use when the state is enabled/active.
-    public var enabledIcon: ComponentStyle
-
-    /// The text style to use when the state is enabled/active.
-    public var enabledText: ComponentStyle
-
-    /// The icon style to use when the state is disabled/inactive.
-    public var disabledIcon: ComponentStyle
-
-    /// The text style to use when the state is disabled/inactive.
-    public var disabledText: ComponentStyle
-}
-
-public extension KeyboardStateLabelStyle {
-
-    /**
-     This is the standard ``KeyboardStateLabel`` style. It
-
-
-     You can modify this style to globally affect every view
-     that is created with the standard style.
-     */
-    static var standard = KeyboardStateLabelStyle()
-}
-
-
-struct KeyboardStateLabel_Previews: PreviewProvider {
-
-    static var previews: some View {
-        List {
-            KeyboardStateLabel(
-                isEnabled: true,
-                enabledText: "Enabled",
-                disabledText: "Disabled"
-            )
-            KeyboardStateLabel(
-                isEnabled: false,
-                enabledText: "Enabled",
-                disabledText: "Disabled"
-            )
+        Section {
+            ForEach([true, false], id: \.self) {
+                KeyboardStateLabel(
+                    isEnabled: $0,
+                    enabledIcon: Color.green,
+                    enabledText: "Yes",
+                    disabledIcon: Color.red,
+                    disabledText: "No"
+                )
+            }
         }
     }
 }