+++ Sources/KeyboardKit/_Pro/ProPlaceholders.swift
@@ -22,47 +22,74 @@ public enum ProPlaceholderError: LocalizedError {
     }
 }
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public struct EmojiKeyboard {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public struct EmojiKeyboardStyle {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public class ExternalKeyboardContext: ObservableObject {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public class iPadProKeyboardLayoutProvider: iPadKeyboardLayoutProvider {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public struct KeyboardTextField {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
-public struct KeyboardTheme {}
+public struct KeyboardTextView {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
+///
+/// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
+public struct KeyboardTheme {
+    
+    public static var allPredefined: [KeyboardTheme] {
+        get throws { throw ProPlaceholderError.proPlaceholder }
+    }
+    
+    public static var standard: KeyboardTheme {
+        get throws { throw ProPlaceholderError.proPlaceholder }
+    }
+    
+    public static var swifty: KeyboardTheme {
+        get throws { throw ProPlaceholderError.proPlaceholder }
+    }
+    
+    public static var minimal: KeyboardTheme {
+        get throws { throw ProPlaceholderError.proPlaceholder }
+    }
+}
+
+/// This is unlocked by KeyboardKit Pro.
+///
+/// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
+public class ThemeBasedKeyboardStyleProvider {}
+
+/// This is unlocked by KeyboardKit Pro.
 /// 
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public class LocalAutocompleteProvider {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public class ProDictationService {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public class ProKeyboardDictationService {
@@ -70,22 +97,22 @@ public class ProKeyboardDictationService {
     public func tryToReturnToKeyboard() {}
 }
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public class RemoteAutocompleteProvider {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public protocol SpeechRecognizer {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public struct SystemKeyboardPreview {}
 
-/// This type is unlocked by KeyboardKit Pro.
+/// This is unlocked by KeyboardKit Pro.
 ///
 /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
 public struct SystemKeyboardButtonPreview {}
@@ -120,7 +147,7 @@ public extension Image {
 
 public extension InputSet {
 
-    /// This type is unlocked by KeyboardKit Pro.
+    /// This is unlocked by KeyboardKit Pro.
     ///
     /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
     static var azerty: InputSet {
@@ -129,7 +156,7 @@ public extension InputSet {
         }
     }
 
-    /// This type is unlocked by KeyboardKit Pro.
+    /// This is unlocked by KeyboardKit Pro.
     ///
     /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
     static var qwertz: InputSet {
@@ -140,14 +167,28 @@ public extension InputSet {
 }
 
 
+public extension KeyboardStyleProvider {
+    
+    /// This is unlocked by KeyboardKit Pro.
+    ///
+    /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
+    static func themed(
+        with theme: KeyboardTheme,
+        context: KeyboardContext
+    ) throws -> Self {
+        throw ProPlaceholderError.proPlaceholder
+    }
+}
+
+
 public extension Proxy {
     
-    /// This type is unlocked by KeyboardKit Pro.
+    /// This is unlocked by KeyboardKit Pro.
     ///
     /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
     struct FullDocumentConfiguration {}
     
-    /// This type is unlocked by KeyboardKit Pro.
+    /// This is unlocked by KeyboardKit Pro.
     ///
     /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
     struct FullDocumentResult {
@@ -163,14 +204,14 @@ public extension Proxy {
 
 public extension View {
     
-    /// This type is unlocked by KeyboardKit Pro.
+    /// This is unlocked by KeyboardKit Pro.
     ///
     /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
     func emojiKeyboardStyle(_ style: EmojiKeyboardStyle) -> some View {
         Text(ProPlaceholderError.proPlaceholder.localizedDescription)
     }
     
-    /// This type is unlocked by KeyboardKit Pro.
+    /// This is unlocked by KeyboardKit Pro.
     ///
     /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
     func keyboardDictation<Overlay: View>() -> some View {
@@ -182,7 +223,7 @@ public extension View {
 #if os(iOS) || os(tvOS) || os(visionOS)
 public extension UITextDocumentProxy {
     
-    /// This type is unlocked by KeyboardKit Pro.
+    /// This is unlocked by KeyboardKit Pro.
     ///
     /// See [KeyboardKit Pro](https://github.com/Keyboard/KeyboardKitPro) for more information.
     func fullDocumentContext(